requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pycparser==2.19
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
six==1.12.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.5
